embedder/requirements.txt
sentence-transformers
psycopg2-binary
python-dotenv
